csv-ingestion/requirements.txt
# Data Warehouse ETL Dependencies
sqlalchemy>=1.4.0
alembic>=1.8.0
psycopg2-binary>=2.9.0
python-dotenv>=0.20.0
pandas>=1.4.0
click>=8.1.0
streamlit>=1.32.0